requirements.txt
aiohttp==3.8.1
aiohttp-apispec==2.2.2
pymysql==1.0.2
sqlalchemy==1.4.31
alembic==1.7.6
passlib==1.7.4
PyJWT==2.3.0
pydantic==1.9.0